Step4 로또(수동) (#3569)
* docs : 기능 목록 작성

* feat : 수동 개수와 번호 입력

* chore : final keyword

* refactor : 정적팩토리메서드로 변경

* chore : 수정

* refactor : LottoNumber 캐싱

* chore : 수정

* refactor : 캐싱 활용

* chore : 자바 코드 컨벤션

* chore : Lotto 생성자 접근제한자 변경

* chore : LottoNumber 생성자 접근제한자 변경

* refactor : Lottos 일급콜렉션으로 변경, test : 테스트 코드 추가

* refactor : price 원시값 포장, test : Money 클래스 테스트 코드

* chore : 자바 코드 컨벤션

+++ src/main/java/lotto/domain/Rank.java
@@ -30,11 +30,11 @@ public enum Rank {
 		return Rank.MISS;
 	}
 
-	public static int calculate(final Rank rank, final int lottoCount) {
+	public static double calculate(final Rank rank, final int lottoCount) {
 		return findPrize(rank) * lottoCount;
 	}
 
-	public static int findPrize(final Rank rank) {
+	public static double findPrize(final Rank rank) {
 		return rank.prize;
 	}
 }
